requirements.txt
aiohttp
appdirs
async-generator
async-timeout
asyncio
attrs
cachetools
certifi
cffi
chardet
cryptography
Cybernator
decorator
discord
distlib
dnspython
embed
filelock
geojson
google-api-core
google-auth
google-auth-httplib2
hentai
HTMLParser
httplib2
idna
importlib-metadata
json-api
linecache2
loli
lolicon
multidict
nekos.py
nhentai
outcome
Pillow
pip
protobuf
pyasn1
pyasn1-modules
pycparser
pymongo
PyNaCl
pyowm
PySocks
python-dotenv
pytz
requests
requests-toolbelt
rsa
setuptools
six
sniffio
sortedcontainers
SpeechRecognition
tabulate
traceback2
trio
typing-extensions
uritemplate
urllib3
validators
virtualenv
virtualenv-clone
vk
websockets
yarl
youtube-dl
zipp